Move all sort qualifier tests into a separate class
